Update realtime-script.md
Your example sends the disconnect message to the player that just disconnected - not the players that are still in the match. I changed it so it would be as intended.

Just a side note - is it possible to change playerId to perhaps playerPeerId? playerId makes me think it's the player's id, not their peer id. Though, getPlayers() does not return playerId so, assuming the variable name playerId is actually the player's peerId is a safe assumption - one that just isn't the easiest to get to if you don't know the system well.